auth/requirements.txt
asgiref==3.8.1
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
cryptography==43.0.0
Django==4.2.16
django-cors-headers==4.5.0
django-otp==1.5.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
importlib_metadata==8.5.0
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
qrcode==8.0
requests==2.32.3
requests-logger==0.2.0
segno==1.6.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
psycopg
